Move duplicate operational days code to one place
