+++ service/src/main/kotlin/fi/espoo/evaka/shared/domain/Time.kt
@@ -1,13 +1,9 @@
-// SPDX-FileCopyrightText: 2017-2020 City of Espoo
+// SPDX-FileCopyrightText: 2017-2022 City of Espoo
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 package fi.espoo.evaka.shared.domain
 
-import fi.espoo.evaka.shared.DaycareId
-import fi.espoo.evaka.shared.db.Database
-import fi.espoo.evaka.shared.db.mapColumn
-import org.jdbi.v3.core.kotlin.mapTo
 import java.time.DayOfWeek
 import java.time.LocalDate
 import java.time.Month
@@ -203,49 +199,6 @@ fun asDistinctPeriods(periods: List<DateRange>, spanningPeriod: DateRange): List
         .sortedBy { orMax(it) }
 
     return allStartDates.map { start -> DateRange(start, allEndDates.find { end -> start <= orMax(end) }) }.toList()
-}
-
-data class OperationalDays(
-    val fullMonth: List<LocalDate>,
-    val generalCase: List<LocalDate>,
-    private val specialCases: Map<DaycareId, List<LocalDate>>
-) {
-    fun forUnit(id: DaycareId): List<LocalDate> = specialCases[id] ?: generalCase
-}
-
-fun Database.Read.operationalDays(year: Int, month: Month): OperationalDays {
-    val range = FiniteDateRange.ofMonth(year, month)
-    val firstDayOfMonth = LocalDate.of(year, month, 1)
-    val daysOfMonth = generateSequence(firstDayOfMonth) { it.plusDays(1) }
-        .takeWhile { date -> date.month == month }
-
-    // Only includes units that don't have regular monday to friday operational days
-    val specialUnitOperationalDays = createQuery("SELECT id, operation_days FROM daycare WHERE NOT (operation_days @> '{1,2,3,4,5}' AND operation_days <@ '{1,2,3,4,5}')")
-        .map { row -> row.mapColumn<DaycareId>("id") to row.mapColumn<Array<Int>>("operation_days").map { DayOfWeek.of(it) } }
-        .toList()
-
-    val holidays = createQuery("SELECT date FROM holiday WHERE between_start_and_end(:range, date)")
-        .bind("range", range)
-        .mapTo<LocalDate>()
-        .list()
-
-    val generalCase = daysOfMonth
-        .filter { it.dayOfWeek != DayOfWeek.SATURDAY && it.dayOfWeek != DayOfWeek.SUNDAY }
-        .filterNot { holidays.contains(it) }
-        .toList()
-
-    val specialCases = specialUnitOperationalDays
-        .map { (unitId, operationalDays) ->
-            val operationalDates = daysOfMonth
-                .filter { operationalDays.contains(it.dayOfWeek) }
-                // Units that are operational every day of the week are also operational during holidays
-                .filter { operationalDays.size == 7 || !holidays.contains(it) }
-
-            unitId to operationalDates.toList()
-        }
-        .toMap()
-
-    return OperationalDays(daysOfMonth.toList(), generalCase, specialCases)
 }
 
 fun LocalDate.isWeekend() = this.dayOfWeek == DayOfWeek.SATURDAY || this.dayOfWeek == DayOfWeek.SUNDAY
